fix(#726): fix GraphiQL configuration
There was some inconsistency in configuration handling. The
`GraphiQLProperties` were not used for headers and GraphiQL props.
Instead, the properties were manually loaded.

This part, however, was not updated with the starter reorganization,
and expected the variables in the old place
(`graphiql.props.variables.`).

This commit fixes this by using the props / headers from the
configuration properties.